moves NotificationSettingsPage into settingPage folder
